Updated snowpack code so no implicit type version is required in variable defintions--also began to implement development of black-box tests for happy jack.

+++ src/tHydro/tSnowPack.cpp
@@ -79,7 +79,7 @@ void tSnowPack::SetSnowPackVariables(tInputFile &infile) {
     densityAge = 0.0;
     ETAge = 0.0;
     compactParam = 0.3;
-    rhoSnFreshkg = 100;
+    rhoSnFreshkg = 100.0;
     snOnOff = 0.0;
 }
 
@@ -91,8 +91,8 @@ void tSnowPack::SetSnowInterceptVariables() {
     kc = 0.010; //-
     iceRad = 500e-6; //m
     Mwater = 18.01; //kg/kmol
-    R = 8313; //J/kmol K
-    RdryAir = 287; //J/kg K
+    R = 8313.0; //J/kmol K
+    RdryAir = 287.0; //J/kg K
     nu = 1.3e-5; //m^2/s
     KtAtm = 0.024; //J/msK WR updated value from Liston and Elder
     esatIce = 0.0;
@@ -106,8 +106,8 @@ void tSnowPack::SetSnowVariables(tInputFile &infile) {
 
     //time steps
     timeStepm = infile.ReadItem(timeStepm, "METSTEP");
-    timeSteph = timeStepm / 60;
-    timeSteps = 60 * timeStepm;
+    timeSteph = timeStepm / 60.0;
+    timeSteps = 60.0 * timeStepm;
     minutelyTimeStep = 0.0;
 
     //state variables
@@ -144,8 +144,8 @@ void tSnowPack::SetSnowVariables(tInputFile &infile) {
     cpicekJ = 2.1;
     cpwaterkJ = 4.190;
     cpairkJ = 1.006;
-    latFreezekJ = 334;
-    latVapkJ = 2470;
+    latFreezekJ = 334.0;
+    latVapkJ = 2470.0;
     latSubkJ = latFreezekJ + latVapkJ;
 
     //output variables
@@ -160,9 +160,9 @@ void tSnowPack::SetSnowVariables(tInputFile &infile) {
     kilotonaught = 1e3;
     cgsRHOtomks = 1e3;
     mksRHOtocgs = 1e-3;
-    naughttocm = 100; // Used convert m to cm
+    naughttocm = 100.0; // Used convert m to cm
     cmtonaught = 0.01; // Used convert cm to m
-    ctom = 10;
+    ctom = 10.0;
     mtoc = 0.1;
 
 }
@@ -584,7 +584,7 @@ void tSnowPack::callSnowPack(tIntercept *Intercept, int flag) {
 
                 //if there is no snow left at this point, then bail out of
                 //energy balance.
-                if ((snWE <= 5e-6) || (snTempC < -800)) {
+                if ((snWE <= 5e-6) || (snTempC < -800.0)) {
                     liqRoute = 0.0;
                     snWE = 0.0;
                     iceWE = 0.0;
@@ -743,7 +743,7 @@ void tSnowPack::callSnowPack(tIntercept *Intercept, int flag) {
         EP = ApproximateEP();
 
         // Submit values to climate simulator
-        weatherSimul->ComputeDailyEpCld(EP, SkyC / cnt);
+        weatherSimul->ComputeDailyEpCld(EP, SkyC / static_cast<double>(cnt));
 
         // Assign the radiation variables to the 'tHydrometStoch'
         if (!count) {
